requirements.txt
argparse
bs4
dnspython
requests
